Change gas amounts to uint64_t.

+++ src/validation.h
@@ -628,7 +628,7 @@ bool LoadMempool();
 //////////////////////////////////////////////////////// qtum
 std::vector<ResultExecute> callContract(const dev::Address& addrContract, std::vector<unsigned char> opcode, const dev::Address& sender = dev::Address());
 
-bool CheckMinGasPrice(std::vector<EthTransactionParams>& etps, const uint32_t& minGasPrice);
+bool CheckMinGasPrice(std::vector<EthTransactionParams>& etps, const uint64_t& minGasPrice);
 
 struct ByteCodeExecResult;
 
@@ -688,7 +688,7 @@ class ByteCodeExec {
 
 public:
 
-    ByteCodeExec(const CBlock& _block, std::vector<QtumTransaction> _txs, const uint32_t _blockGasLimit) : txs(_txs), block(_block), blockGasLimit(_blockGasLimit) {}
+    ByteCodeExec(const CBlock& _block, std::vector<QtumTransaction> _txs, const uint64_t _blockGasLimit) : txs(_txs), block(_block), blockGasLimit(_blockGasLimit) {}
 
     void performByteCode(dev::eth::Permanence type = dev::eth::Permanence::Committed);
 
@@ -708,7 +708,7 @@ private:
 
     const CBlock& block;
 
-    const uint32_t blockGasLimit;
+    const uint64_t blockGasLimit;
 
 };
 ////////////////////////////////////////////////////////
